web-flask/requirements.txt
faiss-cpu==1.12.0
flask-cors==6.0.1
neo4j==5.28.2
openai==2.3.0
opencv-python==4.12.0.88
pdfplumber==0.11.7
pip==25.2
PyJWT==2.10.1
python-docx==1.2.0
requests==2.32.5
setuptools==78.1.1
wheel==0.45.1
